Copy-pasted page structure from pesto.html
Copied half of pesto.html's code starting from
"Description" to the last </ol> tag and inserted it between <img src>
and the closing body tag in the porksinigang.html file.

Afterward, I removed the content inside pasted elements to create
a template to also be pasted in the last recipe to be added.